requirements.txt
pyyaml>=6.0
requests>=2.31
pytest>=8.0
pytest-mock>=3.12
